Move backend code to workspace
