+++ tsconfig.json
@@ -1,9 +1,8 @@
 {
   "compilerOptions": {
-    "target": "ESNext",
     "useDefineForClassFields": true,
-    "module": "CommonJS",
     "sourceMap": true,
+    "strict": true,
     "resolveJsonModule": true,
     "allowSyntheticDefaultImports": true,
     "esModuleInterop": true,
