+++ workspaces/backend/src/types/game-info.ts
