+++ workspaces/backend/src/types/game-state.ts
