+++ workspaces/frontend/src/networking/types.ts
@@ -1,5 +1,5 @@
 import { Socket } from "socket.io-client";
-import { ClientEvents } from "../../backend/src/events/client-events";
-import { ServerEvents } from "../../backend/src/events/server-events";
+import type { ClientEvents } from "../../../backend/src/events/client-events";
+import type { ServerEvents } from "../../../backend/src/events/server-events";
 
 export type TinywarsSocket = Socket<ServerEvents, ClientEvents>;
